code cleanup - moved fetchRecommendations into utils.jsx
